+++ apps/kyb-app/vite.config.ts
@@ -9,7 +9,7 @@ import { defineConfig } from 'vitest/config';
 import checker from 'vite-plugin-checker';
 import tsconfigPaths from 'vite-tsconfig-paths';
 import terminal from 'vite-plugin-terminal';
-
+import topLevelAwait from 'vite-plugin-top-level-await';
 interface PackageJson {
   name: string;
   version: string;
@@ -21,6 +21,10 @@ const packageJson: PackageJson = JSON.parse(
 );
 
 const plugins: PluginOption[] = [
+  topLevelAwait({
+    promiseExportName: '__tla',
+    promiseImportName: i => `__tla_${i}`,
+  }),
   react(),
   tailwindcss(),
   checker({ typescript: true, overlay: false }),