+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/config
 
+## 1.1.25
+
+### Patch Changes
+
+- version bump
+
 ## 1.1.24
 
 ### Patch Changes