+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/websocket-service
 
+## 0.1.25
+
+### Patch Changes
+
+- version bump
+
 ## 0.1.24
 
 ### Patch Changes